+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
 - New command hardis:org:multi-org-query allowing to execute a SOQL Bulk Query in multiple orgs and aggregate the results in a single CSV / XLS report
+- New command hardis:org:community:update to Activate / Deactivate communities from command line
 
 ## [5.3.0] 2024-10-24